Refactor most things
