+++ ankh/main.go
@@ -1,18 +1,15 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
 	"math/rand"
 	"os"
 	"os/signal"
 	"path"
-	"sort"
 	"strconv"
 	"strings"
 	"syscall"
-	"text/tabwriter"
 	"time"
 
 	"github.com/imdario/mergo"
@@ -26,15 +23,19 @@ import (
 	"github.com/appnexus/ankh/context"
 	"github.com/appnexus/ankh/docker"
 	"github.com/appnexus/ankh/helm"
-	"github.com/appnexus/ankh/jira"
 	"github.com/appnexus/ankh/kubectl"
-	"github.com/appnexus/ankh/slack"
 	"github.com/appnexus/ankh/util"
 )
 
 var AnkhBuildVersion string = "DEVELOPMENT"
 
 var log = logrus.New()
+
+func check(err error) {
+	if err != nil {
+		log.Fatalf("%v", err)
+	}
+}
 
 func setLogLevel(ctx *ankh.ExecutionContext, level logrus.Level) {
 	if ctx.Quiet {
@@ -59,650 +60,13 @@ func signalHandler(ctx *ankh.ExecutionContext, sigs chan os.Signal) {
 	}
 }
 
-func printEnvironments(ankhConfig *ankh.AnkhConfig) {
-	keys := []string{}
-	for k, _ := range ankhConfig.Environments {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	for _, name := range keys {
-		log.Infof("* %v", name)
-	}
-}
-
-func getEnvironmentTable(ankhConfig *ankh.AnkhConfig) []string {
-	buf := bytes.NewBufferString("")
-	w := tabwriter.NewWriter(buf, 0, 8, 8, ' ', 0)
-	fmt.Fprintf(w, "NAME\tCONTEXTS\tSOURCE\n")
-	keys := []string{}
-	for k, _ := range ankhConfig.Environments {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	for _, name := range keys {
-		env, _ := ankhConfig.Environments[name]
-		fmt.Fprintf(w, "%v\t%v\t%v\n", name, strings.Join(env.Contexts, ","), env.Source)
-	}
-	w.Flush()
-	return strings.Split(buf.String(), "\n")
-}
-
-func getContextTable(ankhConfig *ankh.AnkhConfig) []string {
-	buf := bytes.NewBufferString("")
-	w := tabwriter.NewWriter(buf, 0, 8, 8, ' ', 0)
-	fmt.Fprintf(w, "NAME\tRELEASE\tENVIRONMENT-CLASS\tRESOURCE-PROFILE\tKUBE-CONTEXT/SERVER\tSOURCE\n")
-	keys := []string{}
-	for k, _ := range ankhConfig.Contexts {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	for _, name := range keys {
-		ctx, _ := ankhConfig.Contexts[name]
-		target := ctx.KubeContext
-		if target == "" {
-			target = ctx.KubeServer
-		}
-		fmt.Fprintf(w, "%v\t%v\t%v\t%v\t%v\t%v\n", name, ctx.Release, ctx.EnvironmentClass, ctx.ResourceProfile, target, ctx.Source)
-	}
-	w.Flush()
-	return strings.Split(buf.String(), "\n")
-}
-
-func printContexts(ankhConfig *ankh.AnkhConfig) {
-	keys := []string{}
-	for k, _ := range ankhConfig.Contexts {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	for _, name := range keys {
-		log.Infof("* %v", name)
-	}
-}
-
-func reconcileMissingConfigs(ctx *ankh.ExecutionContext, ankhFile *ankh.AnkhFile) error {
-	// Make sure that we don't use the tag argument for more than one Chart.
-	// When this happens, it is almost always an error, because a tag value
-	// is typically only valid/intended for a single chart.
-	tagArgumentUsedForChart := ""
-
-	// Prompt for chart versions if any are missing
-	for i := 0; i < len(ankhFile.Charts); i++ {
-		chart := &ankhFile.Charts[i]
-
-		if chart.Path == "" && chart.Version == "" {
-			ctx.Logger.Infof("Found chart \"%v\" without a version", chart.Name)
-			if ctx.NoPrompt {
-				ctx.Logger.Fatalf("Chart \"%v\" missing version (and no 'path' set either, not prompting due to --no-prompt)",
-					chart.Name)
-			}
-
-			repository := ctx.DetermineHelmRepository(&chart.HelmRepository)
-			versions, err := helm.ListVersions(ctx, repository, chart.Name, true)
-			if err != nil {
-				return err
-			}
-
-			selectedVersion, err := util.PromptForSelection(strings.Split(strings.Trim(versions, "\n "), "\n"),
-				fmt.Sprintf("Select a version for chart \"%v\"", chart.Name), false)
-			if err != nil {
-				return err
-			}
-
-			chart.Version = selectedVersion
-			ctx.Logger.Infof("Using chart \"%v\" at version \"%v\" based on prompt selection", chart.Name, chart.Version)
-		} else if chart.Path != "" {
-			ctx.Logger.Infof("Using chart \"%v\" from local path \"%v\"", chart.Name, chart.Path)
-		}
-
-		// Now that we have either a version or a local path, fetch the chart metadata and merge it.
-		repository := ctx.DetermineHelmRepository(&chart.HelmRepository)
-		meta, err := helm.FetchChartMeta(ctx, repository, chart)
-		if err != nil {
-			return fmt.Errorf("Error fetching chart \"%v\": %v", chart.Name, err)
-		}
-		mergo.Merge(&chart.ChartMeta, meta)
-
-		// If namespace is set on the command line, we'll use that as an
-		// override later during executeChartsOnNamespace, so don't check
-		// for anything here.
-		// - command line override, ankh file, chart meta.
-		if ctx.Namespace == nil {
-			if ankhFile.Namespace != nil {
-				extraLog := ""
-				if chart.ChartMeta.Namespace != nil && *ankhFile.Namespace != *chart.ChartMeta.Namespace {
-					extraLog = fmt.Sprintf(" (overriding namespace \"%v\" from ankh.yaml present in the chart)",
-						*chart.ChartMeta.Namespace)
-				}
-				ctx.Logger.Infof("Using namespace \"%v\" from Ankh file for chart \"%v\"%v",
-					*ankhFile.Namespace, chart.Name, extraLog)
-				chart.ChartMeta.Namespace = ankhFile.Namespace
-			} else if chart.ChartMeta.Namespace == nil {
-				ctx.Logger.Infof("Found chart \"%v\" without a namespace", chart.Name)
-				if ctx.NoPrompt {
-					ctx.Logger.Fatalf("Chart \"%v\" missing namespace (not prompting due to --no-prompt)", chart.Name)
-				}
-				if len(ctx.AnkhConfig.Namespaces) > 0 {
-					selectedNamespace, err := util.PromptForSelection(ctx.AnkhConfig.Namespaces,
-						fmt.Sprintf("Select a namespace for chart '%v' (or re-run with -n/--namespace to provide your own)",
-							chart.Name), false)
-					if err != nil {
-						return err
-					}
-					chart.ChartMeta.Namespace = &selectedNamespace
-				} else {
-					providedNamespace, err := util.PromptForInput("",
-						fmt.Sprintf("Provide a namespace for chart '%v' (or enter nothing to denote no explicit namespace) > ",
-							chart.Name))
-					if err != nil {
-						return err
-					}
-					chart.ChartMeta.Namespace = &providedNamespace
-
-				}
-				ctx.Logger.Infof("Using namespace \"%v\" for chart \"%v\" based on prompt selection",
-					*chart.ChartMeta.Namespace, chart.Name)
-			} else {
-				ctx.Logger.Infof("Using namespace \"%v\" for chart \"%v\" based on ankh.yaml present in the chart",
-					*chart.ChartMeta.Namespace, chart.Name)
-			}
-		}
-
-		// tagKey comes directly from the chart's metadata
-		tagKey := chart.ChartMeta.TagKey
-
-		// Do nothing if tagKey is not configured - the user does not want this behavior.
-		if tagKey == "" {
-			if ctx.Tag != nil {
-				ctx.Logger.Fatalf("Tag has been provided but `tagKey` is not configured on either the `chart` in an AnkhFile, nor in an `ankh.yaml` inside the helm chart. " +
-					"This means you passed a tag value, but have not told Ankh which helm value corresponds " +
-					"to the tag value/variable in your helm chart. Tag is shorthand for `--set $tagKey=$tag`, " +
-					"so you can use that instead, or you can ensure that `chart.tagKey` is configured")
-			}
-			continue
-		} else {
-			ctx.Logger.Infof("Using tagKey \"%v\" for chart \"%v\" based on ankh.yaml present in the chart", chart.ChartMeta.TagKey, chart.Name)
-		}
-
-		if ctx.Tag != nil {
-			if tagArgumentUsedForChart != "" {
-				complaint := fmt.Sprintf("Cannot use tag value for chart \"%v\" because it was already used for chart \"%v\". "+
-					"A tag value is almost always intended for use with a single chart. To ignore this error and "+
-					"use tag value \"%v\" for _all_ charts, re-un using `ankh --ignore-config-errors ...` ",
-					chart.Name, tagArgumentUsedForChart, *ctx.Tag)
-				if ctx.IgnoreConfigErrors {
-					ctx.Logger.Warnf(complaint)
-				} else {
-					ctx.Logger.Fatalf(complaint)
-				}
-			}
-
-			ctx.Logger.Infof("Using tag value \"%v=%s\" based on --tag argument", tagKey, *ctx.Tag)
-			chart.Tag = ctx.Tag
-			tagArgumentUsedForChart = chart.Name
-			continue
-		}
-
-		// Treat any existing --set tagKey=$tag argument as authoritative
-		for k, v := range ctx.HelmSetValues {
-			if k == tagKey {
-				ctx.Logger.Infof("Using tag value \"%v=%s\" based on --set argument", tagKey, v)
-				t := v
-				chart.Tag = &t
-				break
-			}
-		}
-
-		// Treat any existing `tag` in `default-values` for this chart as the next-most authoritative
-		for k, v := range chart.DefaultValues {
-			if k == tagKey {
-				ctx.Logger.Infof("Using tag value \"%v=%s\" based on default-values present in the Ankh file", tagKey, v)
-				t, ok := v.(string)
-				if !ok {
-					ctx.Logger.Fatalf("Could not use value '%+v' from default-values in chart %v "+
-						"as a string value for tagKey '%v'", v, chart.Name, tagKey)
-				}
-				chart.Tag = &t
-				break
-			}
-		}
-
-		// For certain operations, we can assume a safe `unset` value for tagKey
-		// for the sole purpose of templating the Helm chart. The value won't be used
-		// meaningfully (like it would be with apply), so we choose this method instead
-		// of prompting the user for a value that isn't meaningful.
-		switch ctx.Mode {
-		case ankh.Rollback:
-			fallthrough
-		case ankh.Get:
-			fallthrough
-		case ankh.Pods:
-			fallthrough
-		case ankh.Exec:
-			fallthrough
-		case ankh.Logs:
-			if chart.Tag != nil {
-				break
-			}
-
-			_, ok := ctx.HelmSetValues[tagKey]
-			if !ok {
-				// It's unset, so set it for the purpose of this execution
-				tag := "__ankh_tag_value_unset___"
-				ctx.Logger.Debugf("Setting configured tagKey %v=%v for a safe operation",
-					tagKey, tag)
-				chart.Tag = &tag
-			}
-		}
-
-		// If we stil don't have a chart.Tag value, prompt.
-		if chart.Tag == nil {
-			if ctx.NoPrompt {
-				ctx.Logger.Fatalf("Chart \"%v\" missing value for `tagKey` (configured to be '%v',  not prompting due to --no-prompt)",
-					tagKey, chart.Name)
-			}
-
-			registryDomain := ctx.AnkhConfig.Docker.Registry
-			image := ""
-			if chart.ChartMeta.TagImage != "" {
-				// No need to prompt for an image name if we already have one in the chart metdata
-				registryDomain, image, err = docker.ParseImage(ctx, chart.ChartMeta.TagImage)
-				check(err)
-
-				ctx.Logger.Infof("Using tagImage \"%v\" for chart \"%v\" based on ankh.yaml present in the chart", chart.ChartMeta.TagImage, chart.Name)
-				ctx.Logger.Debugf("Parsed tagImage into registryDomain '%v' and image '%v'", registryDomain, image)
-			} else {
-				ctx.Logger.Infof("Found chart \"%v\" without a value for \"%v\" ", chart.Name, tagKey)
-				if ctx.AnkhConfig.Docker.Registry == "" {
-					ctx.Logger.Fatalf("Cannot prompt for an image tag, no Docker registry configured.")
-				}
-				defaultValue := chart.Name
-				image, err = util.PromptForInput(defaultValue,
-					fmt.Sprintf("No tag specified for chart '%v'. Provide the name of an image in registry '%v' to select a tag for, "+
-						"or nothing to skip this step > ", ctx.AnkhConfig.Docker.Registry, chart.Name))
-				check(err)
-			}
-
-			if image == "" {
-				ctx.Logger.Infof("Skipping tag prompt since no image name was provided")
-				continue
-			}
-
-			if ctx.AnkhConfig.Docker.Registry == "" {
-				ctx.Logger.Fatalf("Cannot prompt for an image tag, no Docker registry configured.")
-			}
-
-			output, err := docker.ListTags(ctx, registryDomain, image, true)
-			check(err)
-
-			trimmedOutput := strings.Trim(output, "\n ")
-			if trimmedOutput != "" {
-				tags := strings.Split(trimmedOutput, "\n")
-				tag, err := util.PromptForSelection(tags, fmt.Sprintf("Select a value for \"%v\"", tagKey), false)
-				check(err)
-
-				ctx.Logger.Infof("Using implicit \"--set tag %v=%s\" based on prompt selection", tagKey, tag)
-				chart.Tag = &tag
-			} else if image != "" {
-				complaint := fmt.Sprintf("Chart \"%v\" missing value for `tagKey` (configured to be `%v`). "+
-					"You may want to try passing a tag value explicitly using `ankh --set %v=... `, or simply ignore "+
-					"this error entirely using `ankh --ignore-config-errors ...` (not recommended)",
-					chart.Name, tagKey, tagKey)
-				if ctx.IgnoreConfigErrors {
-					ctx.Logger.Warnf("%v", complaint)
-				} else {
-					ctx.Logger.Fatalf("%v", complaint)
-				}
-			}
-		}
-
-		// we should finally have a tag value
-		ctx.DeploymentTag = *chart.Tag
-
-	}
-
-	return nil
-}
-
-func filterOutput(ctx *ankh.ExecutionContext, helmOutput string) string {
-	ctx.Logger.Debugf("Filtering with inclusive list `%v`", ctx.Filters)
-
-	// The golang yaml library doesn't actually support whitespace/comment
-	// preserving round-trip parsing. So, we're going to filter the "hard way".
-	filtered := []string{}
-	objs := strings.Split(helmOutput, "---")
-	for _, obj := range objs {
-		lines := strings.Split(obj, "\n")
-		for _, line := range lines {
-			if !strings.HasPrefix(line, "kind:") {
-				continue
-			}
-			matched := false
-			for _, s := range ctx.Filters {
-				kind := strings.Trim(line[5:], " ")
-				if strings.EqualFold(kind, s) {
-					matched = true
-					break
-				}
-			}
-			if matched {
-				filtered = append(filtered, obj)
-				break
-			}
-		}
-	}
-
-	return "---" + strings.Join(filtered, "---")
-}
-
-func logExecuteAnkhFile(ctx *ankh.ExecutionContext, ankhFile *ankh.AnkhFile) {
-	action := ""
-	switch ctx.Mode {
-	case ankh.Apply:
-		action = "Applying chart"
-	case ankh.Rollback:
-		action = "Rolling back Deployment/StatefulSet from chart"
-	case ankh.Diff:
-		action = "Diffing objects from chart"
-	case ankh.Exec:
-		action = "Exec'ing on pods from chart"
-	case ankh.Explain:
-		action = "Explaining"
-	case ankh.Get:
-		action = "Getting objects from chart"
-	case ankh.Pods:
-		action = "Getting pods for Deployment/StatefulSet from chart"
-	case ankh.Template:
-		action = "Templating"
-	case ankh.Lint:
-		action = "Linting"
-	case ankh.Logs:
-		action = "Getting logs for pods from chart"
-	}
-
-	releaseLog := ""
-	if ctx.AnkhConfig.CurrentContext.Release != "" {
-		releaseLog = fmt.Sprintf(" release \"%v\"", ctx.AnkhConfig.CurrentContext.Release)
-	}
-
-	dryLog := ""
-	if ctx.DryRun {
-		dryLog = " (dry run)"
-	}
-
-	contextLog := fmt.Sprintf(" using kube-context \"%v\"", ctx.AnkhConfig.CurrentContext.KubeContext)
-	if ctx.AnkhConfig.CurrentContext.KubeServer != "" {
-		contextLog = fmt.Sprintf(" to kube-server \"%v\"", ctx.AnkhConfig.CurrentContext.KubeServer)
-	}
-
-	ctx.Logger.Infof("%v%v%v%v with environment class \"%v\" and resource profile \"%v\"", action,
-		releaseLog, dryLog, contextLog,
-		ctx.AnkhConfig.CurrentContext.EnvironmentClass,
-		ctx.AnkhConfig.CurrentContext.ResourceProfile)
-}
-
-func execute(ctx *ankh.ExecutionContext) {
-	rootAnkhFile, err := ankh.GetAnkhFile(ctx)
-	check(err)
-
-	contexts := []string{}
-	if ctx.Environment != "" {
-		environment, ok := ctx.AnkhConfig.Environments[ctx.Environment]
-		if !ok {
-			log.Errorf("Environment '%v' not found in `environments`", ctx.Environment)
-			log.Info("The following environments are available:")
-			printEnvironments(&ctx.AnkhConfig)
-			os.Exit(1)
-		}
-
-		contexts = environment.Contexts
-	}
-
-	if len(contexts) > 0 {
-		log.Infof("Executing over environment \"%v\" with contexts [ %v ]", ctx.Environment, strings.Join(contexts, ", "))
-
-		for _, context := range contexts {
-			log.Infof("Beginning to operate on context \"%v\" in environment \"%v\"", context, ctx.Environment)
-			switchContext(ctx, &ctx.AnkhConfig, context)
-			executeContext(ctx, &rootAnkhFile)
-			log.Infof("Finished with context \"%v\" in environment \"%v\"", context, ctx.Environment)
-		}
-	} else {
-		executeContext(ctx, &rootAnkhFile)
-	}
-
-	if ctx.SlackChannel != "" {
-		if err := slack.PingSlackChannel(ctx); err != nil {
-			ctx.Logger.Errorf("Slack message failed with error: %v", err)
-		}
-	}
-
-	if ctx.CreateJiraTicket {
-		if err := jira.CreateJiraTicket(ctx); err != nil {
-			ctx.Logger.Errorf("Unable to create JIRA ticket. %v", err)
-		}
-	}
-}
-
-func executeChartsOnNamespace(ctx *ankh.ExecutionContext, ankhFile *ankh.AnkhFile,
-	charts []ankh.Chart, namespace string) {
-	// Template, then filter.
-	helmOutput, err := helm.Template(ctx, charts, namespace)
-	check(err)
-	if len(ctx.Filters) > 0 {
-		helmOutput = filterOutput(ctx, helmOutput)
-	}
-
-	// Only pass wildcard labels for "get"-oriented operations.
-	useWildCardLabels := false
-	switch ctx.Mode {
-	case ankh.Diff:
-		fallthrough
-	case ankh.Get:
-		fallthrough
-	case ankh.Pods:
-		fallthrough
-	case ankh.Exec:
-		fallthrough
-	case ankh.Logs:
-		useWildCardLabels = true
-		fallthrough
-	case ankh.Explain:
-		fallthrough
-	case ankh.Rollback:
-		fallthrough
-	case ankh.Apply:
-		if ctx.KubectlVersion == "" {
-			ver, err := kubectl.Version(ctx)
-			if err != nil {
-				ctx.Logger.Fatalf("Failed to get kubectl version info: %v", err)
-			}
-			ctx.KubectlVersion = ver
-			ctx.Logger.Debug("Using kubectl version: ", strings.TrimSpace(ver))
-		}
-
-		// Override wild card labels at the chart level. Choose the first chart arbitrarily.
-		// Warn on this condition - we should eventually deprecate `get/logs/exec` calls
-		// that involve a multi-chart Ankh file.
-		wildCardLabels := []string{}
-		if useWildCardLabels {
-			wildCardLabels = ctx.AnkhConfig.Kubectl.WildCardLabels
-			if charts[0].ChartMeta.WildCardLabels != nil {
-				wildCardLabels = *charts[0].ChartMeta.WildCardLabels
-				ctx.Logger.Debugf("Using override wildCardLabels %+v from chart %v", wildCardLabels, charts[0].Name)
-				if len(ankhFile.Charts) > 1 {
-					ctx.Logger.Warnf("Action \"%v\" over multiple charts will be eventually be deprecated",
-						ctx.Mode)
-				}
-			}
-		}
-
-		kubectlOutput, err := kubectl.Execute(ctx, helmOutput, namespace, wildCardLabels, nil)
-		if err != nil && ctx.Mode == ankh.Diff {
-			ctx.Logger.Warnf("The `diff` feature entered alpha in kubectl v1.9.0, and seems to work best at version v1.12.1. "+
-				"Your results may vary. Current kubectl version string is `%s`", ctx.KubectlVersion)
-		}
-		check(err)
-
-		if ctx.Mode == ankh.Explain {
-			// Sweet string badnesss.
-			helmOutput = strings.TrimSpace(strings.TrimSuffix(strings.TrimSpace(helmOutput), "&& \\"))
-			fmt.Println(fmt.Sprintf("(%s) | \\\n%s", helmOutput, kubectlOutput))
-		} else if kubectlOutput != "" {
-			fmt.Println(kubectlOutput)
-		}
-	case ankh.Template:
-		fmt.Println(helmOutput)
-	case ankh.Lint:
-		errors := helm.Lint(ctx, helmOutput, *ankhFile)
-		if len(errors) > 0 {
-			for _, err := range errors {
-				ctx.Logger.Warningf("%v", err)
-			}
-			log.Fatalf("Lint found %d errors", len(errors))
-		}
-
-		ctx.Logger.Infof("No issues")
-	}
-}
-
-func executeAnkhFile(ctx *ankh.ExecutionContext, ankhFile *ankh.AnkhFile) {
-	err := reconcileMissingConfigs(ctx, ankhFile)
-	check(err)
-
-	logExecuteAnkhFile(ctx, ankhFile)
-
-	if ctx.HelmVersion == "" {
-		ver, err := helm.Version(ctx)
-		if err != nil {
-			ctx.Logger.Fatalf("Failed to get helm version info: %v", err)
-		}
-		ctx.HelmVersion = ver
-		ctx.Logger.Debug("Using helm version: ", strings.TrimSpace(ver))
-	}
-
-	logChartsExecute := func(charts []ankh.Chart, namespace string, extra string) {
-		plural := "s"
-		n := len(charts)
-		if n == 1 {
-			plural = ""
-		}
-		names := []string{}
-		for _, chart := range charts {
-			names = append(names, chart.Name)
-		}
-		ctx.Logger.Infof("Using %vnamespace \"%v\" for %v chart%v [ %v ]",
-			extra, namespace, n, plural, strings.Join(names, ", "))
-	}
-
-	if ctx.Namespace != nil {
-		// Namespace overridden on the command line, so use that one for everything.
-		namespace := *ctx.Namespace
-		logChartsExecute(ankhFile.Charts, namespace, "command-line override ")
-		executeChartsOnNamespace(ctx, ankhFile, ankhFile.Charts, namespace)
-	} else {
-		// Gather charts by namespace, and execute them in sets.
-		chartSets := make(map[string][]ankh.Chart)
-		for _, chart := range ankhFile.Charts {
-			namespace := *chart.ChartMeta.Namespace
-			chartSets[namespace] = append(chartSets[namespace], chart)
-		}
-
-		// Sort the namespaces. We don't guarantee this behavior, but it's more sane than
-		// letting the namespace ordering depend on unorderd golang maps.
-		allNamespaces := []string{}
-		for namespace, _ := range chartSets {
-			allNamespaces = append(allNamespaces, namespace)
-		}
-		sort.Strings(allNamespaces)
-		for _, namespace := range allNamespaces {
-			charts := chartSets[namespace]
-			logChartsExecute(charts, namespace, "")
-			executeChartsOnNamespace(ctx, ankhFile, charts, namespace)
-		}
-	}
-}
-
-func executeContext(ctx *ankh.ExecutionContext, rootAnkhFile *ankh.AnkhFile) {
-	dependencies := []string{}
-	if ctx.Chart == "" {
-		dependencies = rootAnkhFile.Dependencies
-	} else {
-		log.Debugf("Skipping dependencies since we are operating only on chart %v", ctx.Chart)
-	}
-
-	for _, dep := range dependencies {
-		log.Infof("Satisfying dependency: %v", dep)
-
-		ankhFilePath := dep
-		ankhFile, err := ankh.ParseAnkhFile(ankhFilePath)
-		if err == nil {
-			ctx.Logger.Debugf("- OK: %v", ankhFilePath)
-		}
-		check(err)
-
-		ctx.WorkingPath = path.Dir(ankhFilePath)
-		executeAnkhFile(ctx, &ankhFile)
-		ctx.WorkingPath = ""
-
-		log.Infof("Finished satisfying dependency: %v", dep)
-	}
-
-	if len(rootAnkhFile.Charts) > 0 {
-		executeAnkhFile(ctx, rootAnkhFile)
-	} else if len(dependencies) == 0 {
-		if ctx.AnkhConfig.Helm.Repository == "" || ctx.NoPrompt {
-			ctx.Logger.Fatalf("No charts nor dependencies provided, nothing to do")
-		} else {
-			// Prompt for a chart
-			ctx.Logger.Infof("No chart specified as an argument, and no `charts` found in an Ankh file")
-			charts, err := helm.GetChartNames(ctx, ctx.AnkhConfig.Helm.Repository)
-			check(err)
-
-			selectedChart, err := util.PromptForSelection(charts, "Select a chart", false)
-			check(err)
-
-			rootAnkhFile.Charts = []ankh.Chart{ankh.Chart{Name: selectedChart}}
-			ctx.Logger.Infof("Using chart \"%v\" based on prompt selection", selectedChart)
-
-			executeAnkhFile(ctx, rootAnkhFile)
-		}
-	}
-}
-
-func checkContext(ankhConfig *ankh.AnkhConfig, context string) {
-	_, ok := ankhConfig.Contexts[context]
-	if !ok {
-		log.Errorf("Context '%v' not found in `contexts`", context)
-		log.Info("The following contexts are available:")
-		printContexts(ankhConfig)
-		os.Exit(1)
-	}
-}
-
-func switchContext(ctx *ankh.ExecutionContext, ankhConfig *ankh.AnkhConfig, context string) {
-	if context == "" {
-		log.Errorf("No context or environment provided. Provide one using -c/--context or -e/--environment")
-		log.Info("The following contexts are available:")
-		printContexts(ankhConfig)
-		log.Info("The following environments are available:")
-		printEnvironments(ankhConfig)
-		os.Exit(1)
-	}
-
-	checkContext(ankhConfig, context)
-	errs := ankhConfig.ValidateAndInit(ctx, context)
-	if len(errs) > 0 && !ctx.IgnoreContextAndEnv {
-		// The config validation errors are not recoverable.
-		log.Fatalf("%v", util.MultiErrorFormat(errs))
-	}
-}
-
 func main() {
 	app := cli.App("ankh", "Another Kubernetes Helper")
-	app.Spec = "[--verbose] [--quiet] [--no-prompt] [--ignore-config-errors] [--ankhconfig] [--kubeconfig] [--datadir] [--release] [--context] [--environment] [--namespace] [--tag] [--set...]"
+	app.Spec = "[--verbose] [--explain] [--quiet] [--no-prompt] [--ignore-config-errors] [--ankhconfig] [--kubeconfig] [--datadir] [--release] [--context] [--environment] [--namespace] [--tag] [--set...]"
 
 	var (
 		verbose            = app.BoolOpt("v verbose", false, "Verbose debug mode")
+		explain            = app.BoolOpt("explain", false, "Explain mode")
 		quiet              = app.BoolOpt("q quiet", false, "Quiet mode. Critical logging only. The quiet option overrides the verbose option.")
 		noPrompt           = app.BoolOpt("no-prompt", false, "Do not prompt for missing required configuration. Exit with non-zero status and a fatal log message instead.")
 		ignoreConfigErrors = app.BoolOpt("ignore-config-errors", false, "Ignore certain configuration errors that have defined, but potentially dangerous behavior.")
@@ -799,6 +163,7 @@ func main() {
 
 		ctx = &ankh.ExecutionContext{
 			Verbose:             *verbose,
+			Explain:             *explain,
 			Quiet:               *quiet,
 			AnkhConfigPath:      *ankhconfig,
 			KubeConfigPath:      *kubeconfig,
@@ -934,27 +299,6 @@ func main() {
 		ctx.AnkhConfig = mergedAnkhConfig
 	}
 
-	app.Command("explain", "Explain how an Ankh file would be applied to a Kubernetes cluster", func(cmd *cli.Cmd) {
-		cmd.Spec = "[-f] [--chart] [--chart-path]"
-
-		ankhFilePath := cmd.StringOpt("f filename", "ankh.yaml", "Config file name")
-		chart := cmd.StringOpt("chart", "", "Limits the explain command to only the specified chart")
-		chartPath := cmd.StringOpt("chart-path", "", "Use a local chart directory instead of a remote, versioned chart")
-
-		cmd.Action = func() {
-			ctx.AnkhFilePath = *ankhFilePath
-			ctx.Chart = *chart
-			if *chartPath != "" {
-				ctx.Chart = *chartPath
-				ctx.LocalChart = true
-			}
-			ctx.Mode = ankh.Explain
-
-			execute(ctx)
-			os.Exit(0)
-		}
-	})
-
 	app.Command("apply", "Apply an Ankh file to a Kubernetes cluster", func(cmd *cli.Cmd) {
 		cmd.Spec = "[-f] [--dry-run] [--chart] [--chart-path] [--slack] [--slack-message] [--jira-ticket] [--filter...]"
 
@@ -976,6 +320,41 @@ func main() {
 				ctx.LocalChart = true
 			}
 			ctx.Mode = ankh.Apply
+			ctx.SlackChannel = *slackChannel
+			ctx.SlackMessageOverride = *slackMessageOverride
+			ctx.CreateJiraTicket = *createJiraTicket
+			filters := []string{}
+			for _, filter := range *filter {
+				filters = append(filters, string(filter))
+			}
+			ctx.Filters = filters
+
+			execute(ctx)
+			os.Exit(0)
+		}
+	})
+
+	app.Command("deploy", "Run a multi-stage deployment of an Ankh file to a Kubernetes cluster", func(cmd *cli.Cmd) {
+		cmd.Spec = "[-f] [--dry-run] [--chart] [--chart-path] [--slack] [--slack-message] [--jira-ticket] [--filter...]"
+
+		ankhFilePath := cmd.StringOpt("f filename", "ankh.yaml", "Config file name")
+		dryRun := cmd.BoolOpt("dry-run", false, "Perform a dry-run and don't actually deploy anything to a cluster")
+		chart := cmd.StringOpt("chart", "", "Limits the deploy command to only the specified chart")
+		chartPath := cmd.StringOpt("chart-path", "", "Use a local chart directory instead of a remote, versioned chart")
+		slackChannel := cmd.StringOpt("s slack", "", "Send slack message to specified slack channel about application update")
+		slackMessageOverride := cmd.StringOpt("m slack-message", "", "Override the default slack message being sent")
+		createJiraTicket := cmd.BoolOpt("j jira-ticket", false, "Create a JIRA ticket to track update")
+		filter := cmd.StringsOpt("filter", []string{}, "Kubernetes object kinds to include for the action. The entries in this list are case insensitive. Any object whose `kind:` does not match this filter will be excluded from the action")
+
+		cmd.Action = func() {
+			ctx.AnkhFilePath = *ankhFilePath
+			ctx.DryRun = *dryRun
+			ctx.Chart = *chart
+			if *chartPath != "" {
+				ctx.Chart = *chartPath
+				ctx.LocalChart = true
+			}
+			ctx.Mode = ankh.Deploy
 			ctx.SlackChannel = *slackChannel
 			ctx.SlackMessageOverride = *slackMessageOverride
 			ctx.CreateJiraTicket = *createJiraTicket
@@ -1013,7 +392,6 @@ func main() {
 			ctx.SlackChannel = *slackChannel
 			ctx.SlackMessageOverride = *slackMessageOverride
 			ctx.CreateJiraTicket = *createJiraTicket
-			ctx.Filters = []string{"deployment", "statfulset"}
 
 			ctx.Logger.Warnf("Rollback is not a transactional operation.\n" +
 				"\n" +
@@ -1398,9 +776,10 @@ func main() {
 	app.Command("config", "Manage Ankh configuration", func(cmd *cli.Cmd) {
 		ctx.IgnoreContextAndEnv = true
 		ctx.IgnoreConfigErrors = true
-		ctx.SkipConfig = true
 
 		cmd.Command("init", "Initialize Ankh configuration", func(cmd *cli.Cmd) {
+			ctx.SkipConfig = true
+
 			cmd.Action = func() {
 				// Use the original, unmerged config. We want to explicitly avoid
 				// serializing the contents of any remote configs.
@@ -1452,6 +831,8 @@ func main() {
 		})
 
 		cmd.Command("add", "Add an Ankh configuration source", func(cmd *cli.Cmd) {
+			ctx.SkipConfig = true
+
 			cmd.Spec = "SOURCE"
 			sourceArg := cmd.StringArg("SOURCE", "", "The configuration source to add. May be a local file or a remote HTTP resource.")
 
@@ -1488,6 +869,8 @@ func main() {
 		})
 
 		cmd.Command("rm", "Remove an Ankh configuration source", func(cmd *cli.Cmd) {
+			ctx.SkipConfig = true
+
 			cmd.Spec = "SOURCE"
 			sourceArg := cmd.StringArg("SOURCE", "", "The configuration source to remove. May be a local file or a remote HTTP resource.")
 
@@ -1578,10 +961,4 @@ func main() {
 	})
 
 	app.Run(os.Args)
-}
-
-func check(err error) {
-	if err != nil {
-		log.Fatalf("%v", err)
-	}
 }
