--- helm/helm_test.go
@@ -1,8 +1,0 @@
-package helm
-
-import (
-	"testing"
-)
-
-// TODO: write tests
-func TestStub(t *testing.T) {}
